pyproject.toml
[project]
name = "llm-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=0.130.0",
    "fastmcp>=2.8.0",
    "google-generativeai>=0.8.5",
    "langchain>=0.3.25",
    "langchain-community>=0.3.25",
    "openai>=1.86.0",
    "streamlit>=1.45.1",
]
